Node Wrangler: refactor by splitting the script into several files
After discussion in !104444 and !104457 it became apparent that Node
Wrangler needs a bit of work to make its maintenance more comfortable.
So, split the code into interface, operators and preferences, and a
few specific utility files.

The bulk of the add-on now resides in operators.py.
